+++ tests/test_handlers_llm_evaluation.py
@@ -1,6 +1,7 @@
-"""Tests for effectful.handlers.llm.type_checking."""
+"""Tests for effectful.handlers.llm.evaluation."""
 
 import ast
+import builtins
 import inspect
 import textwrap
 import types
@@ -12,8 +13,11 @@ from typing import Annotated, Any, TypedDict
 
 import pydantic
 import pytest
+from RestrictedPython import RestrictingNodeTransformer
 
+from effectful.handlers.llm.encoding import Encodable, SynthesizedFunction
 from effectful.handlers.llm.evaluation import (
+    RestrictedEvalProvider,
     collect_imports,
     collect_runtime_type_stubs,
     collect_variable_declarations,
@@ -21,6 +25,7 @@ from effectful.handlers.llm.evaluation import (
     type_to_ast,
 )
 from effectful.internals.unification import nested_type
+from effectful.ops.semantics import handler
 from effectful.ops.syntax import defop
 from effectful.ops.types import NotHandled
 
@@ -65,12 +70,12 @@ def get_context() -> Mapping[str, Any]:
         if not (
             (
                 isinstance(v, types.ModuleType)
-                and v.__name__.startswith("tests.test_handlers_llm_type_checking")
+                and v.__name__.startswith("tests.test_handlers_llm_evaluation")
             )
             or (
                 hasattr(v, "__module__")
                 and (
-                    v.__module__.startswith("tests.test_handlers_llm_type_checking")
+                    v.__module__.startswith("tests.test_handlers_llm_evaluation")
                     or v.__module__.startswith("_pytest")
                 )
                 and inspect.isclass(v)
@@ -1387,3 +1392,96 @@ class TestMypyTypeCheckNameCollision:
         module = ast.parse(source)
         ctx = get_context()
         mypy_type_check(module, ctx, [int], bool)
+
+
+# ============================================================================
+# RestrictedEvalProvider security tests
+# ============================================================================
+
+
+def test_restricted_blocks_private_attribute_access():
+    """RestrictedPython blocks access to underscore-prefixed attributes by default."""
+    encodable = Encodable.define(Callable[[str], int], {})
+    source = SynthesizedFunction(
+        module_code="""def get_private(s: str) -> int:
+    return s.__class__.__name__"""
+    )
+    # Should raise due to restricted attribute access
+    with pytest.raises(Exception):  # Could be NameError or AttributeError
+        with handler(RestrictedEvalProvider()):
+            fn = encodable.decode(source)
+            fn("test")
+
+
+def test_restricted_with_custom_policy():
+    """Can pass custom policy via kwargs."""
+
+    # Create a custom policy that's the same as default (just to test the plumbing)
+    class CustomPolicy(RestrictingNodeTransformer):
+        pass
+
+    encodable = Encodable.define(Callable[[int, int], int], {})
+    source = SynthesizedFunction(
+        module_code="""def add(a: int, b: int) -> int:
+    return a + b"""
+    )
+    with handler(RestrictedEvalProvider(policy=CustomPolicy)):
+        fn = encodable.decode(source)
+    assert fn(2, 3) == 5
+
+
+def test_builtins_in_env_does_not_bypass_security():
+    """Including __builtins__ in env should not bypass RestrictedEvalProvider security.
+
+    RestrictedEvalProvider explicitly filters out __builtins__ from the env
+    to prevent callers from replacing the restricted builtins with full Python builtins.
+    This test verifies that even if __builtins__ is passed in the context,
+    dangerous operations remain blocked.
+    """
+
+    # Attempt to pass full builtins in the context, which should be filtered out
+    dangerous_ctx = {"__builtins__": builtins.__dict__}
+
+    # Test 1: open() should not be usable even with __builtins__ in context
+    encodable_open = Encodable.define(Callable[[str], str], dangerous_ctx)
+    source_open = SynthesizedFunction(
+        module_code="""def read_file(path: str) -> str:
+    return open(path).read()"""
+    )
+    with pytest.raises(Exception):  # Could be NameError, ValueError, or other
+        with handler(RestrictedEvalProvider()):
+            fn = encodable_open.decode(source_open)
+            fn("/etc/passwd")
+
+    # Test 2: __import__ should not be usable
+    encodable_import = Encodable.define(Callable[[], str], dangerous_ctx)
+    source_import = SynthesizedFunction(
+        module_code="""def get_os_name() -> str:
+    os = __import__('os')
+    return os.name"""
+    )
+    with pytest.raises(Exception):
+        with handler(RestrictedEvalProvider()):
+            fn = encodable_import.decode(source_import)
+            fn()
+
+    # Test 3: Verify safe code still works with dangerous context
+    encodable_safe = Encodable.define(Callable[[int, int], int], dangerous_ctx)
+    source_safe = SynthesizedFunction(
+        module_code="""def add(a: int, b: int) -> int:
+    return a + b"""
+    )
+    with handler(RestrictedEvalProvider()):
+        fn = encodable_safe.decode(source_safe)
+        assert fn(2, 3) == 5, "Safe code should still work"
+
+    # Test 4: Private attribute access should still be blocked
+    encodable_private = Encodable.define(Callable[[str], str], dangerous_ctx)
+    source_private = SynthesizedFunction(
+        module_code="""def get_class(s: str) -> str:
+    return s.__class__.__name__"""
+    )
+    with pytest.raises(Exception):
+        with handler(RestrictedEvalProvider()):
+            fn = encodable_private.decode(source_private)
+            fn("test")
